Create a driver interface and move DB behind MySQL flavor
