+++ server/lib/request_context.js
@@ -2,7 +2,9 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-var App = require('../models/app');
+var requireDriver = require('../lib/db').requireDriver;
+
+var App = requireDriver('../models', 'app');
 
 module.exports = function(cb) {
   return function(req, res) {
@@ -31,7 +33,6 @@ module.exports = function(cb) {
         return cb(req, res, ctx);
       });
     } else {
-
       return cb(req, res, ctx);
     }
   };
